# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "전동기 API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/checks": {
      "post": {
        "tags": [
          "계약 전 확인"
        ],
        "summary": "Create Check",
        "operationId": "create_check_api_checks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResult"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/checks/{check_id}": {
      "get": {
        "tags": [
          "계약 전 확인"
        ],
        "summary": "Get Check",
        "operationId": "get_check_api_checks__check_id__get",
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Check Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/registry/parse": {
      "post": {
        "tags": [
          "등기부 점검"
        ],
        "summary": "Parse Registry",
        "operationId": "parse_registry_api_registry_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_registry_api_registry_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryParseResult"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_parse_registry_api_registry_parse_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_registry_api_registry_parse_post"
      },
      "CheckInput": {
        "properties": {
          "housing_type": {
            "type": "string",
            "enum": [
              "apartment",
              "row_house",
              "officetel",
              "multi_household",
              "detached"
            ],
            "title": "Housing Type"
          },
          "deposit": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Deposit",
            "description": "보증금 (원)"
          },
          "official_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Official Price",
            "description": "공동주택 공시가격 (원)"
          },
          "trade_price": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Price",
            "description": "같은 단지·비슷한 면적 매매 실거래가 (원)"
          },
          "exclusive_area": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclusive Area",
            "description": "전용면적 (㎡)"
          },
          "registry": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegistryInput"
              },
              {
                "type": "null"
              }
            ],
            "description": "등기부를 올리기 전이면 비워 둠"
          },
          "trust_consent": {
            "type": "string",
            "enum": [
              "unknown",
              "obtained",
              "unavailable"
            ],
            "title": "Trust Consent",
            "default": "unknown"
          },
          "building_violation": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Violation",
            "description": "건축물대장 위반건축물 여부, 모르면 비워 둠"
          },
          "guarantor_consult": {
            "type": "string",
            "enum": [
              "not_yet",
              "possible",
              "impossible"
            ],
            "title": "Guarantor Consult",
            "default": "not_yet"
          }
        },
        "type": "object",
        "required": [
          "housing_type",
          "deposit"
        ],
        "title": "CheckInput",
        "description": "계약 전 확인 입력 (FR-001)."
      },
      "CheckResult": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "next_step": {
            "type": "string",
            "title": "Next Step"
          },
          "block_payment": {
            "type": "boolean",
            "title": "Block Payment"
          },
          "estimated_price": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Price"
          },
          "deposit_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Ratio"
          },
          "ratio_is_minimum": {
            "type": "boolean",
            "title": "Ratio Is Minimum"
          },
          "matched": {
            "items": {
              "$ref": "#/components/schemas/MatchedRuleOut"
            },
            "type": "array",
            "title": "Matched"
          },
          "linkages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linkages"
          },
          "unchecked": {
            "items": {
              "$ref": "#/components/schemas/UncheckedItemOut"
            },
            "type": "array",
            "title": "Unchecked"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          },
          "rules_version": {
            "type": "string",
            "title": "Rules Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "level",
          "label",
          "next_step",
          "block_payment",
          "estimated_price",
          "deposit_ratio",
          "ratio_is_minimum",
          "matched",
          "linkages",
          "unchecked",
          "disclaimer",
          "rules_version"
        ],
        "title": "CheckResult"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "화면 분기용 고정 코드 (docs/api/API_명세.md 1.5절)"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "사용자에게 보여 줄 문장"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint",
            "description": "다음에 할 일 안내"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ErrorField"
            },
            "type": "array",
            "title": "Fields",
            "description": "입력 형식 오류일 때 필드별 문제"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorField": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorField"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "모든 오류 응답의 형식."
      },
      "MatchedRuleOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "linkage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkage"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "level"
        ],
        "title": "MatchedRuleOut"
      },
      "RegistryEntryOut": {
        "properties": {
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "purpose",
          "text"
        ],
        "title": "RegistryEntryOut"
      },
      "RegistryFieldsOut": {
        "properties": {
          "seizure": {
            "type": "boolean",
            "title": "Seizure"
          },
          "mortgage": {
            "type": "boolean",
            "title": "Mortgage"
          },
          "mortgage_amount": {
            "type": "integer",
            "title": "Mortgage Amount"
          },
          "trust": {
            "type": "boolean",
            "title": "Trust"
          }
        },
        "type": "object",
        "required": [
          "seizure",
          "mortgage",
          "mortgage_amount",
          "trust"
        ],
        "title": "RegistryFieldsOut",
        "description": "RegistryInput과 같은 이름. 채권최고액을 못 읽으면 mortgage=true, mortgage_amount=0으로 나갈 수 있어 검사는 하지 않는다."
      },
      "RegistryInput": {
        "properties": {
          "seizure": {
            "type": "boolean",
            "title": "Seizure",
            "default": false
          },
          "mortgage": {
            "type": "boolean",
            "title": "Mortgage",
            "default": false
          },
          "mortgage_amount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Mortgage Amount",
            "description": "근저당 채권최고액 합계 (원)",
            "default": 0
          },
          "trust": {
            "type": "boolean",
            "title": "Trust",
            "default": false
          }
        },
        "type": "object",
        "title": "RegistryInput",
        "description": "등기부 \"주요 등기사항 요약\" 추출 결과 (사용자 확인값, FR-005)."
      },
      "RegistryMortgageOut": {
        "properties": {
          "rank": {
            "type": "string",
            "title": "Rank"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "holder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holder"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "purpose",
          "text",
          "amount",
          "holder"
        ],
        "title": "RegistryMortgageOut"
      },
      "RegistryParseResult": {
        "properties": {
          "fields": {
            "$ref": "#/components/schemas/RegistryFieldsOut"
          },
          "owners": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Owners"
          },
          "mortgages": {
            "items": {
              "$ref": "#/components/schemas/RegistryMortgageOut"
            },
            "type": "array",
            "title": "Mortgages"
          },
          "seizures": {
            "items": {
              "$ref": "#/components/schemas/RegistryEntryOut"
            },
            "type": "array",
            "title": "Seizures"
          },
          "trusts": {
            "items": {
              "$ref": "#/components/schemas/RegistryEntryOut"
            },
            "type": "array",
            "title": "Trusts"
          },
          "others": {
            "items": {
              "$ref": "#/components/schemas/RegistryEntryOut"
            },
            "type": "array",
            "title": "Others"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "fields",
          "owners",
          "mortgages",
          "seizures",
          "trusts",
          "others",
          "warnings"
        ],
        "title": "RegistryParseResult",
        "description": "등기부 요약 추출 결과 (FR-005). 사용자가 확인·수정한 뒤 fields를 CheckInput.registry로 보낸다 (FR-006)."
      },
      "UncheckedItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "linkage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkage"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label"
        ],
        "title": "UncheckedItemOut"
      }
    }
  }
}